Moved equals into children of ArrayValue
I did not particularly like this change, but there seems to be no good
way that would make the checkstyle rule 'EqualsHashCode' happy
("all classes that overwrite one of those methods, must also overwrite
the other" -> but there is no good way to overwrite hashCode in
ArrayValue because inheritance)
