+++ community/values/src/main/java/org/neo4j/values/storable/BooleanArray.java
@@ -22,6 +22,7 @@ package org.neo4j.values.storable;
 import java.util.Arrays;
 
 import org.neo4j.values.AnyValue;
+import org.neo4j.values.SequenceValue;
 
 import static java.lang.String.format;
 
@@ -146,6 +147,21 @@ abstract class BooleanArray extends ArrayValue
     public String prettyPrint()
     {
         return Arrays.toString( value() );
+    }
+
+    @Override
+    public final boolean equals( Object other )
+    {
+        if ( other == null )
+        {
+            return false;
+        }
+
+        if ( other instanceof SequenceValue )
+        {
+            return this.equals( (SequenceValue) other );
+        }
+        return other instanceof Value && equals( (Value) other );
     }
 
     @Override
