+++ community/values/src/main/java/org/neo4j/values/storable/FloatArray.java
@@ -22,6 +22,7 @@ package org.neo4j.values.storable;
 import java.util.Arrays;
 
 import org.neo4j.values.AnyValue;
+import org.neo4j.values.SequenceValue;
 
 import static java.lang.String.format;
 
@@ -87,6 +88,21 @@ abstract class FloatArray extends FloatingPointArray
     public boolean equals( double[] x )
     {
         return PrimitiveArrayValues.equals( value(), x  );
+    }
+
+    @Override
+    public final boolean equals( Object other )
+    {
+        if ( other == null )
+        {
+            return false;
+        }
+
+        if ( other instanceof SequenceValue )
+        {
+            return this.equals( (SequenceValue) other );
+        }
+        return other instanceof Value && equals( (Value) other );
     }
 
     @Override
